Prototype/server/requirements.txt
flask
flask-cors
tensorflow
pillow
numpy
requests
python-dotenv
supabase
matplotlib